% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_haplotype_id_to_haplotype.R
\name{convert_haplotype_id_to_haplotype}
\alias{convert_haplotype_id_to_haplotype}
\title{Convert a haplotype ID to a haplotype}
\usage{
convert_haplotype_id_to_haplotype(haplotype_id)
}
\arguments{
\item{haplotype_id}{the \code{nmhc2ppreds} haplotype ID,
for example \code{h1}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{nmhc2ppreds} haplotype ID to an netmhc2pan haplotype name}
}
\description{
Convert a haplotype ID to a hapolotype,
in netmhc2pan naming format
}
\examples{
convert_haplotype_id_to_haplotype("h13")
}
\author{
Richèl J.C. Bilderbeek
}
