% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_haplotype_to_haplotype_id.R
\name{convert_haplotype_to_haplotype_id}
\alias{convert_haplotype_to_haplotype_id}
\title{Convert a haplotype to a haplotype ID}
\usage{
convert_haplotype_to_haplotype_id(haplotype)
}
\arguments{
\item{haplotype}{the netmhc2pan haplotype ID,
for example \code{HLA-A33:01}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{nmhc2ppreds} haplotype ID to an netmhc2pan haplotype name}
}
\description{
Convert a haplotype (in netmhc2pan naming format) to a hapolotype ID
}
\examples{
convert_haplotype_to_haplotype_id("DRB1_0311")
}
\author{
Richèl J.C. Bilderbeek
}
