% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_haplotypes_to_haplotype_ids.R
\name{convert_haplotypes_to_haplotype_ids}
\alias{convert_haplotypes_to_haplotype_ids}
\title{Convert haplotypes to haplotype IDs}
\usage{
convert_haplotypes_to_haplotype_ids(haplotypes)
}
\arguments{
\item{haplotypes}{one ore more netmhc2pan haplotype IDs,
for example \code{HLA-A33:01}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{nmhc2ppreds} haplotype ID to an netmhc2pan haplotype name}
}
\description{
Convert one or more haplotypes (in netmhc2pan naming format)
to haplotype IDs
}
\examples{
convert_haplotypes_to_haplotype_ids("HLA-A-33:01")
}
\author{
Richèl J.C. Bilderbeek
}
