% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ic50s_to_quantiles.R
\name{convert_ic50s_to_quantiles}
\alias{convert_ic50s_to_quantiles}
\title{Convert IC50s to a quantile table}
\usage{
convert_ic50s_to_quantiles(ic50s, n = 100)
}
\arguments{
\item{ic50s}{the IC50 values}

\item{n}{the number of steps from 0.0 to 1.0.
For example, use 0.1 for 10 steps}
}
\value{
a tibble with two columns: \code{q} the quantiles,
and \code{ic50} the IC50s
}
\description{
Convert IC50s to a quantile table
}
