% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_quantile_to_ic50.R
\name{convert_quantile_to_ic50}
\alias{convert_quantile_to_ic50}
\title{Convert a quantile to an IC50}
\usage{
convert_quantile_to_ic50(lut, q)
}
\arguments{
\item{lut}{lookup table}

\item{q}{the quantile}
}
\description{
Convert a quantile to an IC50
}
