% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_quantile_to_ic50.R
\name{convert_quantile_to_ic50_new}
\alias{convert_quantile_to_ic50_new}
\title{Convert a quantile to an IC50, new school}
\usage{
convert_quantile_to_ic50_new(lut, q)
}
\arguments{
\item{lut}{lookup table}

\item{q}{the quantile}
}
\description{
Convert a quantile to an IC50, new school
}
