% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_peptide.R
\name{create_random_peptide}
\alias{create_random_peptide}
\title{Create a random peptide}
\usage{
create_random_peptide(length)
}
\arguments{
\item{length}{length of the peptide in amino acids}
}
\description{
Create a random peptide
}
\examples{
create_random_peptide()
}
\author{
Richèl J.C. Bilderbeek
}
