% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lut.R
\name{get_lut}
\alias{get_lut}
\title{Get the lookup-table}
\usage{
get_lut(peptide_length, mhc_haplotype)
}
\arguments{
\item{peptide_length}{length of the peptide in amino acids}

\item{mhc_haplotype}{haplotype of either HMC-I or MHC-II.
in netmhc2panr format.
Use \code{netmhc2panr::get_mhc_1_haplotypes}
to get a list of all MHC-I haplotypes.
Use \code{netmhc2panr::get_mhc_2_haplotypes}
to get a list of all MHC-II haplotypes.}
}
\description{
Get the lookup-table for a specific
peptide size and haplotype.
}
